+++ backend/src/main/java/team/teamby/teambyteam/feed/application/dto/FeedThreadWritingRequest.java
@@ -1,9 +1,10 @@
 package team.teamby.teambyteam.feed.application.dto;
 
+import org.springframework.web.multipart.MultipartFile;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import org.springframework.web.multipart.MultipartFile;
 
 public record FeedThreadWritingRequest(
         String content,
