+++ backend/src/main/java/team/teamby/teambyteam/feed/domain/FeedRepository.java
@@ -1,7 +1,6 @@
 package team.teamby.teambyteam.feed.domain;
 
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
