+++ backend/src/main/java/team/teamby/teambyteam/feed/presentation/FeedThreadController.java
@@ -1,7 +1,6 @@
 package team.teamby.teambyteam.feed.presentation;
 
 import jakarta.validation.Valid;
-import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,18 +10,22 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import team.teamby.teambyteam.feed.application.FeedThreadService;
+import team.teamby.teambyteam.feed.application.FeedReadService;
+import team.teamby.teambyteam.feed.application.FeedWriteService;
 import team.teamby.teambyteam.feed.application.dto.FeedThreadWritingRequest;
 import team.teamby.teambyteam.feed.application.dto.FeedsResponse;
 import team.teamby.teambyteam.member.configuration.AuthPrincipal;
 import team.teamby.teambyteam.member.configuration.dto.MemberEmailDto;
+
+import java.net.URI;
 
 @RestController
 @RequestMapping("/api/team-place")
 @RequiredArgsConstructor
 public class FeedThreadController {
 
-    private final FeedThreadService feedThreadService;
+    private final FeedReadService feedReadService;
+    private final FeedWriteService feedWriteService;
 
     @PostMapping("/{teamPlaceId}/feed/threads")
     public ResponseEntity<Void> write(
@@ -30,7 +33,7 @@ public class FeedThreadController {
             @AuthPrincipal final MemberEmailDto memberEmailDto,
             @PathVariable final Long teamPlaceId
     ) {
-        final Long threadId = feedThreadService.write(request, memberEmailDto, teamPlaceId);
+        final Long threadId = feedWriteService.write(request, memberEmailDto, teamPlaceId);
         final URI location = URI.create("/api/team-place/" + teamPlaceId + "/feed/threads/" + threadId);
 
         return ResponseEntity.created(location).build();
@@ -42,7 +45,7 @@ public class FeedThreadController {
             @AuthPrincipal final MemberEmailDto memberEmailDto,
             @RequestParam final Integer size
     ) {
-        FeedsResponse feeds = feedThreadService.firstRead(teamPlaceId, memberEmailDto, size);
+        FeedsResponse feeds = feedReadService.firstRead(teamPlaceId, memberEmailDto, size);
 
         return ResponseEntity.ok(feeds);
     }
@@ -54,7 +57,7 @@ public class FeedThreadController {
             @RequestParam("last-thread-id") final Long threadId,
             @RequestParam final Integer size
     ) {
-        FeedsResponse feeds = feedThreadService.reRead(teamPlaceId, memberEmailDto, threadId, size);
+        FeedsResponse feeds = feedReadService.reRead(teamPlaceId, memberEmailDto, threadId, size);
 
         return ResponseEntity.ok(feeds);
     }
