+++ backend/src/test/java/team/teamby/teambyteam/feed/acceptance/FeedThreadAcceptanceTest.java
@@ -7,19 +7,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.multipart.MultipartFile;
 import team.teamby.teambyteam.common.AcceptanceTest;
 import team.teamby.teambyteam.common.fixtures.FeedThreadFixtures;
-import team.teamby.teambyteam.common.fixtures.FeedThreadImageFixtures;
 import team.teamby.teambyteam.common.fixtures.MemberFixtures;
 import team.teamby.teambyteam.feed.application.dto.FeedResponse;
 import team.teamby.teambyteam.feed.application.dto.FeedsResponse;
 import team.teamby.teambyteam.feed.domain.Feed;
 import team.teamby.teambyteam.feed.domain.FeedThread;
-import team.teamby.teambyteam.feed.domain.cache.RecentFeedCache;
 import team.teamby.teambyteam.feed.domain.notification.schedulenotification.ScheduleNotification;
 import team.teamby.teambyteam.feed.domain.vo.Content;
 import team.teamby.teambyteam.filesystem.FileStorageManager;
@@ -30,10 +27,7 @@ import team.teamby.teambyteam.schedule.domain.vo.Span;
 import team.teamby.teambyteam.schedule.domain.vo.Title;
 import team.teamby.teambyteam.teamplace.domain.TeamPlace;
 
-import java.time.Clock;
-import java.time.Instant;
 import java.time.LocalDateTime;
-import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -62,18 +56,6 @@ public class FeedThreadAcceptanceTest extends AcceptanceTest {
 
     @MockBean
     private FileStorageManager fileStorageManager;
-
-    @MockBean
-    private RecentFeedCache recentFeedCache;
-
-    @SpyBean
-    private Clock clock;
-
-    @BeforeEach
-    void setup() {
-        given(recentFeedCache.isCached(any(Long.class), any(int.class)))
-                .willReturn(false);
-    }
 
     @Nested
     @DisplayName("피드에 스레드 등록 시")
@@ -299,33 +281,6 @@ public class FeedThreadAcceptanceTest extends AcceptanceTest {
                 softly.assertThat(feedResponse1.images().size()).isEqualTo(2);
                 softly.assertThat(feedResponse2.images().size()).isEqualTo(2);
                 softly.assertThat(feedResponse3.images().size()).isEqualTo(0);
-            });
-        }
-
-        @Test
-        @DisplayName("이미지가 만료되는 경우에 조회한다.")
-        void readIfImageExpired() {
-            // given
-            POST_FEED_THREAD_ONLY_IMAGE_REQUEST(authToken, participatedTeamPlace.getId(), List.of(UNDER_SIZE_PNG_FILE1, UNDER_SIZE_PNG_FILE2));
-
-            final Instant expiredDate = LocalDateTime.now().plusDays(FeedThreadImageFixtures.IMAGE_EXPIRATION_DATE).plusNanos(1).toInstant(ZoneOffset.systemDefault().getRules().getOffset(LocalDateTime.now()));
-            given(clock.instant()).willReturn(expiredDate);
-
-            final Long teamPlaceId = participatedMemberTeamPlace.getId();
-            final int size = 5;
-
-            // when
-            final ExtractableResponse<Response> response = GET_FEED_THREAD_FIRST(authToken, teamPlaceId, size);
-            final FeedsResponse feedsResponse = response.as(FeedsResponse.class);
-            final FeedResponse feedResponse = feedsResponse.threads().get(0);
-
-            //then
-            assertSoftly(softly -> {
-                softly.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-                softly.assertThat(feedsResponse.threads().size()).isEqualTo(1);
-                softly.assertThat(feedResponse.images().size()).isEqualTo(2);
-                softly.assertThat(feedResponse.images().get(0).isExpired()).isTrue();
-                softly.assertThat(feedResponse.images().get(1).isExpired()).isTrue();
             });
         }
