+++ backend/src/test/java/team/teamby/teambyteam/feed/application/FeedReadServiceTest.java
@@ -1,22 +1,33 @@
 package team.teamby.teambyteam.feed.application;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-import static team.teamby.teambyteam.common.fixtures.FeedThreadFixtures.CONTENT_AND_IMAGE_REQUEST;
-import static team.teamby.teambyteam.common.fixtures.FeedThreadFixtures.CONTENT_ONLY_REQUEST;
-import static team.teamby.teambyteam.common.fixtures.FeedThreadFixtures.EMPTY_REQUEST;
-import static team.teamby.teambyteam.common.fixtures.FeedThreadFixtures.IMAGE_ONLY_REQUEST;
-import static team.teamby.teambyteam.common.fixtures.FeedThreadFixtures.NOT_ALLOWED_IMAGE_EXTENSION_REQUEST;
-import static team.teamby.teambyteam.common.fixtures.FeedThreadFixtures.OVER_IMAGE_COUNT_REQUEST;
-import static team.teamby.teambyteam.common.fixtures.FeedThreadFixtures.OVER_IMAGE_SIZE_REQUEST;
-import static team.teamby.teambyteam.common.fixtures.MemberFixtures.PHILIP;
-import static team.teamby.teambyteam.common.fixtures.MemberFixtures.PHILIP_EMAIL;
-import static team.teamby.teambyteam.common.fixtures.MemberFixtures.ROY;
-import static team.teamby.teambyteam.common.fixtures.MemberTeamPlaceFixtures.PHILIP_ENGLISH_TEAM_PLACE;
-import static team.teamby.teambyteam.common.fixtures.TeamPlaceFixtures.ENGLISH_TEAM_PLACE;
-import static team.teamby.teambyteam.common.fixtures.TeamPlaceFixtures.JAPANESE_TEAM_PLACE;
+import org.assertj.core.api.SoftAssertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.mockito.BDDMockito;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.SpyBean;
+import team.teamby.teambyteam.common.ServiceTest;
+import team.teamby.teambyteam.common.fixtures.FeedThreadFixtures;
+import team.teamby.teambyteam.common.fixtures.FeedThreadImageFixtures;
+import team.teamby.teambyteam.feed.application.dto.FeedResponse;
+import team.teamby.teambyteam.feed.application.dto.FeedsResponse;
+import team.teamby.teambyteam.feed.domain.Feed;
+import team.teamby.teambyteam.feed.domain.FeedThread;
+import team.teamby.teambyteam.feed.domain.FeedType;
+import team.teamby.teambyteam.feed.domain.image.FeedThreadImage;
+import team.teamby.teambyteam.feed.domain.image.vo.ImageName;
+import team.teamby.teambyteam.feed.domain.image.vo.ImageUrl;
+import team.teamby.teambyteam.feed.domain.notification.schedulenotification.ScheduleNotification;
+import team.teamby.teambyteam.feed.domain.vo.Content;
+import team.teamby.teambyteam.member.configuration.dto.MemberEmailDto;
+import team.teamby.teambyteam.member.domain.Member;
+import team.teamby.teambyteam.member.domain.MemberTeamPlace;
+import team.teamby.teambyteam.member.domain.vo.DisplayMemberName;
+import team.teamby.teambyteam.schedule.application.event.ScheduleCreateEvent;
+import team.teamby.teambyteam.schedule.domain.vo.Span;
+import team.teamby.teambyteam.schedule.domain.vo.Title;
+import team.teamby.teambyteam.teamplace.domain.TeamPlace;
 
 import java.time.Clock;
 import java.time.Instant;
@@ -24,175 +35,20 @@ import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Stream;
 
-import org.assertj.core.api.SoftAssertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.mock.mockito.SpyBean;
-import org.springframework.web.multipart.MultipartFile;
-import team.teamby.teambyteam.common.ServiceTest;
-import team.teamby.teambyteam.common.fixtures.FeedThreadFixtures;
-import team.teamby.teambyteam.common.fixtures.FeedThreadImageFixtures;
-import team.teamby.teambyteam.feed.application.dto.FeedResponse;
-import team.teamby.teambyteam.feed.application.dto.FeedThreadWritingRequest;
-import team.teamby.teambyteam.feed.application.dto.FeedsResponse;
-import team.teamby.teambyteam.feed.domain.Feed;
-import team.teamby.teambyteam.feed.domain.FeedThread;
-import team.teamby.teambyteam.feed.domain.FeedType;
-import team.teamby.teambyteam.feed.domain.cache.RecentFeedCache;
-import team.teamby.teambyteam.feed.domain.notification.schedulenotification.ScheduleNotification;
-import team.teamby.teambyteam.feed.domain.vo.Content;
-import team.teamby.teambyteam.feed.exception.FeedException;
-import team.teamby.teambyteam.filesystem.FileStorageManager;
-import team.teamby.teambyteam.member.configuration.dto.MemberEmailDto;
-import team.teamby.teambyteam.member.domain.Member;
-import team.teamby.teambyteam.member.domain.MemberTeamPlace;
-import team.teamby.teambyteam.member.domain.vo.DisplayMemberName;
-import team.teamby.teambyteam.member.exception.MemberException;
-import team.teamby.teambyteam.schedule.application.event.ScheduleCreateEvent;
-import team.teamby.teambyteam.schedule.domain.vo.Span;
-import team.teamby.teambyteam.schedule.domain.vo.Title;
-import team.teamby.teambyteam.teamplace.domain.TeamPlace;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+import static team.teamby.teambyteam.common.fixtures.MemberFixtures.PHILIP;
+import static team.teamby.teambyteam.common.fixtures.MemberFixtures.PHILIP_EMAIL;
+import static team.teamby.teambyteam.common.fixtures.MemberFixtures.ROY;
+import static team.teamby.teambyteam.common.fixtures.MemberTeamPlaceFixtures.PHILIP_ENGLISH_TEAM_PLACE;
+import static team.teamby.teambyteam.common.fixtures.TeamPlaceFixtures.ENGLISH_TEAM_PLACE;
+import static team.teamby.teambyteam.common.fixtures.TeamPlaceFixtures.JAPANESE_TEAM_PLACE;
 
-class FeedThreadServiceTest extends ServiceTest {
+class FeedReadServiceTest extends ServiceTest {
 
     @Autowired
-    private FeedThreadService feedThreadService;
-
-    @MockBean
-    private FileStorageManager fileStorageManager;
-
-    @MockBean
-    private RecentFeedCache recentFeedCache;
-
-    @SpyBean
-    private Clock clock;
-
-    @BeforeEach
-    void invalidateCaching() {
-        given(recentFeedCache.isCached(any(Long.class), any(int.class)))
-                .willReturn(false);
-
-    }
-
-    @Nested
-    @DisplayName("피드에 스레드 작성시")
-    class WriteThread {
-
-        @BeforeEach
-        void setup() {
-            given(fileStorageManager.upload(any(MultipartFile.class), any(String.class), any(String.class)))
-                    .willReturn("https://s3://seongha-seeik");
-        }
-
-        static Stream<FeedThreadWritingRequest> requests() {
-            return Stream.of(
-                    CONTENT_ONLY_REQUEST,
-                    IMAGE_ONLY_REQUEST,
-                    CONTENT_AND_IMAGE_REQUEST
-            );
-        }
-
-        @ParameterizedTest
-        @MethodSource("requests")
-        @DisplayName("피드에 스레드를 작성한다.")
-        void writeThreadSuccess(final FeedThreadWritingRequest request) {
-            // given
-            final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
-            final Member author = testFixtureBuilder.buildMember(PHILIP());
-            testFixtureBuilder.buildMemberTeamPlace(author, teamPlace);
-
-            // when
-            final Long feedId = feedThreadService.write(request, new MemberEmailDto(author.getEmail().getValue()),
-                    teamPlace.getId());
-
-            //then
-            assertThat(feedId).isNotNull();
-        }
-
-        @Test
-        @DisplayName("이미지 개수가 4개보다 많으면 예외가 발생한다.")
-        void failWhenOverImageCount() {
-            // given
-            final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
-            final Member author = testFixtureBuilder.buildMember(PHILIP());
-            final FeedThreadWritingRequest request = OVER_IMAGE_COUNT_REQUEST;
-
-            // when & then
-            assertThatThrownBy(() -> feedThreadService.write(request, new MemberEmailDto(author.getEmail().getValue()),
-                    teamPlace.getId()))
-                    .isInstanceOf(FeedException.ImageOverCountException.class)
-                    .hasMessageContaining("허용된 이미지의 개수를 초과했습니다.");
-        }
-
-        @Test
-        @DisplayName("이미지 크기가 허용된 크기보다 많으면 예외가 발생한다.")
-        void failWhenOverImageSize() {
-            // given
-            final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
-            final Member author = testFixtureBuilder.buildMember(PHILIP());
-            final FeedThreadWritingRequest request = OVER_IMAGE_SIZE_REQUEST;
-
-            // when & then
-            assertThatThrownBy(() -> feedThreadService.write(request, new MemberEmailDto(author.getEmail().getValue()),
-                    teamPlace.getId()))
-                    .isInstanceOf(FeedException.ImageSizeException.class)
-                    .hasMessageContaining("허용된 이미지의 크기를 초과했습니다.");
-        }
-
-        @Test
-        @DisplayName("이미지의 확장자가 허용되지 않은 확장자면 예외가 발생한다.")
-        void failWhenNotAllowedImageExtension() {
-            // given
-            final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
-            final Member author = testFixtureBuilder.buildMember(PHILIP());
-            final FeedThreadWritingRequest request = NOT_ALLOWED_IMAGE_EXTENSION_REQUEST;
-
-            // when & then
-            assertThatThrownBy(() -> feedThreadService.write(request, new MemberEmailDto(author.getEmail().getValue()),
-                    teamPlace.getId()))
-                    .isInstanceOf(FeedException.NotAllowedImageExtensionException.class)
-                    .hasMessageContaining("허용되지 않은 확장자입니다.");
-        }
-
-        @Test
-        @DisplayName("내용과 이미지가 모두 존재하지 않으면 예외가 발생한다.")
-        void failWhenContentAndImageNotExist() {
-            // given
-            final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
-            final Member author = testFixtureBuilder.buildMember(PHILIP());
-            final FeedThreadWritingRequest request = EMPTY_REQUEST;
-
-            // when & then
-            assertThatThrownBy(() -> feedThreadService.write(request, new MemberEmailDto(author.getEmail().getValue()),
-                    teamPlace.getId()))
-                    .isInstanceOf(FeedException.WritingRequestEmptyException.class)
-                    .hasMessageContaining("내용과 이미지가 모두 존재하지 않습니다.");
-        }
-
-        @Test
-        @DisplayName("존재하지 않는 멤버로 요청을 보내게 되면 예외가 발생한다.")
-        void failUnAuthorized() {
-            // given
-            final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
-            final Member author = testFixtureBuilder.buildMember(PHILIP());
-            final FeedThreadWritingRequest request = CONTENT_AND_IMAGE_REQUEST;
-
-            // when & then
-            assertThatThrownBy(
-                    () -> feedThreadService.write(request, new MemberEmailDto(author.getEmail().getValue() + "x"),
-                            teamPlace.getId()))
-                    .isInstanceOf(MemberException.MemberNotFoundException.class)
-                    .hasMessageContaining("조회한 멤버가 존재하지 않습니다.");
-        }
-    }
+    private FeedReadService feedReadService;
 
     @Nested
     @DisplayName("피드의 스레드 조회시")
@@ -207,7 +63,7 @@ class FeedThreadServiceTest extends ServiceTest {
             final MemberEmailDto memberEmailDto = new MemberEmailDto(PHILIP_EMAIL);
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), memberEmailDto, size);
 
             //then
             assertThat(feedsResponse.threads()).isEmpty();
@@ -233,7 +89,7 @@ class FeedThreadServiceTest extends ServiceTest {
             final int size = 10;
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), memberEmailDto, size);
             final FeedResponse contentOnly = feedsResponse.threads().get(0);
             final FeedResponse imageOnly = feedsResponse.threads().get(1);
             final FeedResponse contentAndImage = feedsResponse.threads().get(2);
@@ -262,17 +118,17 @@ class FeedThreadServiceTest extends ServiceTest {
             final MemberEmailDto memberEmailDto = new MemberEmailDto(member.getEmailValue());
             final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
             testFixtureBuilder.buildMemberTeamPlace(member, teamPlace);
+            final FeedThreadImage EXPIRED_IMAGE = BDDMockito.spy(new FeedThreadImage(new ImageUrl("expired.com"), new ImageName("expiredImageName")));
+            BDDMockito.given(EXPIRED_IMAGE.isExpired())
+                    .willReturn(true);
 
             final Feed feed1 = testFixtureBuilder.buildFeed(FeedThreadFixtures.CONTENT_AND_IMAGE(teamPlace.getId(), member.getId()));
-            testFixtureBuilder.buildFeedThreadFeedThreadImage((FeedThread) feed1, FeedThreadImageFixtures.A_FEED_THREAD_IMAGE);
-
-            final Instant expiredDate = LocalDateTime.now().plusDays(FeedThreadImageFixtures.IMAGE_EXPIRATION_DATE).plusNanos(1).toInstant(ZoneOffset.systemDefault().getRules().getOffset(LocalDateTime.now()));
-            given(clock.instant()).willReturn(expiredDate);
+            testFixtureBuilder.buildFeedThreadFeedThreadImage((FeedThread) feed1, EXPIRED_IMAGE);
 
             final int size = 10;
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), memberEmailDto, size);
             final FeedResponse contentAndImage = feedsResponse.threads().get(0);
 
             //then
@@ -281,36 +137,6 @@ class FeedThreadServiceTest extends ServiceTest {
                 softly.assertThat(contentAndImage.content()).isEqualTo(feed1.getContent().getValue());
                 softly.assertThat(contentAndImage.images()).isNotEmpty();
                 softly.assertThat(contentAndImage.images().get(0).isExpired()).isTrue();
-            });
-        }
-
-        @Test
-        @DisplayName("이미지는 90일이 지나지 않으면 만료되지 않는다.")
-        void imageNotExpireBefore90Days() {
-            // given
-            final Member member = testFixtureBuilder.buildMember(PHILIP());
-            final MemberEmailDto memberEmailDto = new MemberEmailDto(member.getEmailValue());
-            final TeamPlace teamPlace = testFixtureBuilder.buildTeamPlace(ENGLISH_TEAM_PLACE());
-            testFixtureBuilder.buildMemberTeamPlace(member, teamPlace);
-
-            final Feed feed1 = testFixtureBuilder.buildFeed(FeedThreadFixtures.CONTENT_AND_IMAGE(teamPlace.getId(), member.getId()));
-            testFixtureBuilder.buildFeedThreadFeedThreadImage((FeedThread) feed1, FeedThreadImageFixtures.A_FEED_THREAD_IMAGE);
-
-            final Instant expiredDate = LocalDateTime.now().plusNanos(1).toInstant(ZoneOffset.systemDefault().getRules().getOffset(LocalDateTime.now()));
-            given(clock.instant()).willReturn(expiredDate);
-
-            final int size = 10;
-
-            // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
-            final FeedResponse contentAndImage = feedsResponse.threads().get(0);
-
-            //then
-            SoftAssertions.assertSoftly(softly -> {
-                softly.assertThat(contentAndImage.id()).isEqualTo(1);
-                softly.assertThat(contentAndImage.content()).isEqualTo(feed1.getContent().getValue());
-                softly.assertThat(contentAndImage.images()).isNotEmpty();
-                softly.assertThat(contentAndImage.images().get(0).isExpired()).isFalse();
             });
         }
 
@@ -332,7 +158,7 @@ class FeedThreadServiceTest extends ServiceTest {
             final int size = 10;
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), PHILIP_EMAIL_DTO, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), PHILIP_EMAIL_DTO, size);
             feedsResponse.threads().forEach(System.out::println);
 
             // then
@@ -362,7 +188,7 @@ class FeedThreadServiceTest extends ServiceTest {
             final int size = 10;
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(PHILIP_ENGLISH_TEAM_PLACE.getId(),
+            final FeedsResponse feedsResponse = feedReadService.firstRead(PHILIP_ENGLISH_TEAM_PLACE.getId(),
                     memberEmailDto, size);
 
             //then
@@ -391,7 +217,7 @@ class FeedThreadServiceTest extends ServiceTest {
             testFixtureBuilder.buildFeeds(feeds);
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), memberEmailDto, size);
 
             //then
             SoftAssertions.assertSoftly(softly -> {
@@ -416,7 +242,7 @@ class FeedThreadServiceTest extends ServiceTest {
             final String type = FeedType.THREAD.name().toLowerCase();
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), memberEmailDto, size);
 
             //then
             SoftAssertions.assertSoftly(softly -> {
@@ -438,7 +264,7 @@ class FeedThreadServiceTest extends ServiceTest {
             final MemberEmailDto memberEmailDto = new MemberEmailDto(PHILIP_EMAIL);
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), memberEmailDto, size);
 
             //then
             SoftAssertions.assertSoftly(softly -> {
@@ -468,7 +294,7 @@ class FeedThreadServiceTest extends ServiceTest {
             final int size = 5;
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), memberEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), memberEmailDto, size);
 
             //then
             SoftAssertions.assertSoftly(softly -> {
@@ -498,7 +324,7 @@ class FeedThreadServiceTest extends ServiceTest {
             testFixtureBuilder.buildFeeds(feeds);
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.reRead(teamPlace.getId(),
+            final FeedsResponse feedsResponse = feedReadService.reRead(teamPlace.getId(),
                     new MemberEmailDto(member.getEmailValue()), 4L, size);
 
             //then
@@ -530,7 +356,7 @@ class FeedThreadServiceTest extends ServiceTest {
             testFixtureBuilder.buildFeeds(feeds);
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.reRead(englishTeamPlace.getId(),
+            final FeedsResponse feedsResponse = feedReadService.reRead(englishTeamPlace.getId(),
                     new MemberEmailDto(member.getEmailValue()), 5L, size);
 
             //then
@@ -553,7 +379,7 @@ class FeedThreadServiceTest extends ServiceTest {
             testFixtureBuilder.buildFeeds(feeds);
 
             // when
-            final FeedsResponse feedsResponse = feedThreadService.firstRead(teamPlace.getId(), philipEmailDto, size);
+            final FeedsResponse feedsResponse = feedReadService.firstRead(teamPlace.getId(), philipEmailDto, size);
 
             // then
             SoftAssertions.assertSoftly(softly -> {
