+++ backend/src/test/java/team/teamby/teambyteam/feed/domain/vo/ContentTest.java
@@ -49,5 +49,4 @@ class ContentTest {
         );
     }
 
-
 }
